pyproject.toml
[project]
name = "quant-alpha-foundation"
version = "0.1.0"
description = "A Data Engineering Zoomcamp-inspired quantitative alpha research pipeline."
requires-python = ">=3.10"
dependencies = [
    "fastavro>=1.9.4",
    "duckdb>=0.10.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "plotly>=5.20.0",
    "pyarrow>=15.0.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "streamlit>=1.32.0",
    "typer>=0.12.0",
    "yfinance>=0.2.37",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]
dbt = [
    "dbt-core>=1.8.0",
    "dbt-duckdb>=1.8.0",
    "dbt-bigquery>=1.8.0",
]
ml = [
    "scikit-learn>=1.4.0",
]
batch = [
    "pyspark>=3.5.0",
]
streaming = [
    "confluent-kafka>=2.5.0",
]
dlt = [
    "dlt[duckdb]>=1.0.0",
]
cloud = [
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-storage>=2.18.0",
]

[project.scripts]
quant-alpha = "quant_alpha.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
